Lambda functions are now easier to read. Plus, some refactoring
